requirements.txt
alabaster==0.7.12
Babel==2.9.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
cryptography==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2.10
django-allauth==0.44.0
django-appconf==1.0.4
django-debug-toolbar==3.2
django-extensions==3.1.0
django-filter==2.2.0
django-import-export==1.1.0
django-select2==7.5.0
django-simple-history==2.8.0
django-widget-tweaks==1.4.8
docutils==0.16
edx-sphinx-theme==1.5.0
idna==2.10
imagesize==1.2.0
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.7
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycparser==2.20
Pygments==2.7.3
PyJWT==1.7.1
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.3.1
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.2
